# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WorkingStyle = (typeof WorkingStyle)[keyof typeof WorkingStyle];

export const WorkingStyle = {
  quiet: "quiet",
  light_chat: "light_chat",
  brief_social: "brief_social",
} as const;

export type WorkIntent = (typeof WorkIntent)[keyof typeof WorkIntent];

export const WorkIntent = {
  deep_work: "deep_work",
  studying: "studying",
  job_search: "job_search",
  side_projects: "side_projects",
} as const;

export type UserProfileMaxTravelMinutes =
  (typeof UserProfileMaxTravelMinutes)[keyof typeof UserProfileMaxTravelMinutes];

export const UserProfileMaxTravelMinutes = {
  NUMBER_10: 10,
  NUMBER_20: 20,
  NUMBER_30: 30,
} as const;

export interface UserProfile {
  id: string;
  userId: string;
  name: string;
  photoUrl?: string | null;
  workingStyle: WorkingStyle;
  preferredNeighborhoods: string[];
  maxTravelMinutes: UserProfileMaxTravelMinutes;
  workIntents: WorkIntent[];
  linkedinUrl?: string | null;
  instagramUrl?: string | null;
  hidesSocialLinks: boolean;
  showsOnlyToMatches: boolean;
  reliabilityScore: number;
  checkInCount: number;
  createdAt: string;
  updatedAt: string;
}

export type UpsertProfileInputMaxTravelMinutes =
  (typeof UpsertProfileInputMaxTravelMinutes)[keyof typeof UpsertProfileInputMaxTravelMinutes];

export const UpsertProfileInputMaxTravelMinutes = {
  NUMBER_10: 10,
  NUMBER_20: 20,
  NUMBER_30: 30,
} as const;

export interface UpsertProfileInput {
  name: string;
  photoUrl?: string | null;
  workingStyle: WorkingStyle;
  preferredNeighborhoods: string[];
  maxTravelMinutes: UpsertProfileInputMaxTravelMinutes;
  workIntents: WorkIntent[];
  linkedinUrl?: string | null;
  instagramUrl?: string | null;
  hidesSocialLinks: boolean;
  showsOnlyToMatches: boolean;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Match {
  id: string;
  userId: string;
  matchedUserId: string;
  matchedUserProfile: UserProfile;
  sessionTimeSlot: string;
  sessionDate: string;
  status: MatchStatus;
  weekRevealedAt: string;
  chatEnabled: boolean;
}

export interface ChatMessage {
  id: string;
  matchId: string;
  senderId: string;
  senderName: string;
  content: string;
  sentAt: string;
}

export interface SendMessageInput {
  content: string;
  senderId: string;
  senderName: string;
}

export interface CafeSuggestion {
  id: string;
  name: string;
  neighborhood: string;
  address: string;
  distance: string;
  vibe: string;
}