# lib/api-zod/src/generated/types/cafeSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CafeSuggestion {
  id: string;
  name: string;
  neighborhood: string;
  address: string;
  distance: string;
  vibe: string;
}