# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatus } from "./matchStatus";
import type { UserProfile } from "./userProfile";

export interface Match {
  id: string;
  userId: string;
  matchedUserId: string;
  matchedUserProfile: UserProfile;
  sessionTimeSlot: string;
  sessionDate: Date;
  status: MatchStatus;
  weekRevealedAt: Date;
  chatEnabled: boolean;
}