# lib/api-zod/src/generated/types/sendMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageInput {
  content: string;
  senderId: string;
  senderName: string;
}