# lib/api-zod/src/generated/types/upsertProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpsertProfileInputMaxTravelMinutes } from "./upsertProfileInputMaxTravelMinutes";
import type { WorkingStyle } from "./workingStyle";
import type { WorkIntent } from "./workIntent";

export interface UpsertProfileInput {
  name: string;
  photoUrl?: string | null;
  workingStyle: WorkingStyle;
  preferredNeighborhoods: string[];
  maxTravelMinutes: UpsertProfileInputMaxTravelMinutes;
  workIntents: WorkIntent[];
  linkedinUrl?: string | null;
  instagramUrl?: string | null;
  hidesSocialLinks: boolean;
  showsOnlyToMatches: boolean;
}