# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileMaxTravelMinutes } from "./userProfileMaxTravelMinutes";
import type { WorkingStyle } from "./workingStyle";
import type { WorkIntent } from "./workIntent";

export interface UserProfile {
  id: string;
  userId: string;
  name: string;
  photoUrl?: string | null;
  workingStyle: WorkingStyle;
  preferredNeighborhoods: string[];
  maxTravelMinutes: UserProfileMaxTravelMinutes;
  workIntents: WorkIntent[];
  linkedinUrl?: string | null;
  instagramUrl?: string | null;
  hidesSocialLinks: boolean;
  showsOnlyToMatches: boolean;
  reliabilityScore: number;
  checkInCount: number;
  createdAt: Date;
  updatedAt: Date;
}