# lib/api-zod/src/generated/types/workIntent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */

export type WorkIntent = (typeof WorkIntent)[keyof typeof WorkIntent];

export const WorkIntent = {
  deep_work: "deep_work",
  studying: "studying",
  job_search: "job_search",
  side_projects: "side_projects",
} as const;