# lib/api-zod/src/generated/types/workingStyle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkMate API specification
 * OpenAPI spec version: 0.1.0
 */

export type WorkingStyle = (typeof WorkingStyle)[keyof typeof WorkingStyle];

export const WorkingStyle = {
  quiet: "quiet",
  light_chat: "light_chat",
  brief_social: "brief_social",
} as const;